pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kgs_pipeline"
version = "0.1.0"
description = "KGS Oil & Gas Production Data Pipeline"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "pandas>=2.1.0",
    "pyarrow>=14.0.0",
    "dask[dataframe]>=2024.1.0",
    "distributed>=2024.1.0",
    "bokeh>=3.0.0",
    "numpy>=1.26.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "pandas-stubs>=2.1.0",
    "types-requests>=2.31.0",
    "types-PyYAML>=6.0.0",
    "types-beautifulsoup4>=4.12.0",
]

[project.scripts]
kgs-pipeline = "kgs_pipeline.pipeline:_cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["kgs_pipeline*"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
warn_return_any = false
warn_unused_ignores = false
